+++ app/src/main/AndroidManifest.xml
@@ -65,12 +65,12 @@
             android:name=".view.earn.WebTaskActivity"
             android:screenOrientation="landscape" />
 
-        <service android:name=".network.firebase.KinInstanceIdService">
+        <service android:name=".firebase.KinInstanceIdService">
             <intent-filter>
                 <action android:name="com.google.firebase.INSTANCE_ID_EVENT" />
             </intent-filter>
         </service>
-        <service android:name=".network.firebase.KinMessagingService">
+        <service android:name=".firebase.KinMessagingService">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
             </intent-filter>
