+++ app/src/main/java/org/kinecosystem/kinit/KinitApplication.kt
@@ -7,8 +7,8 @@ import com.squareup.picasso.Picasso
 import io.fabric.sdk.android.Fabric
 import org.kinecosystem.kinit.analytics.Analytics
 import org.kinecosystem.kinit.dagger.*
-import org.kinecosystem.kinit.network.ServicesProvider
-import org.kinecosystem.kinit.network.Wallet
+import org.kinecosystem.kinit.server.ServicesProvider
+import org.kinecosystem.kinit.blockchain.Wallet
 import org.kinecosystem.kinit.repository.DataStore
 import org.kinecosystem.kinit.repository.DataStoreProvider
 import org.kinecosystem.kinit.repository.SharedPreferencesStore
