+++ app/src/main/java/org/kinecosystem/kinit/firebase/KinInstanceIdService.java
@@ -1,10 +1,10 @@
-package org.kinecosystem.kinit.network.firebase;
+package org.kinecosystem.kinit.firebase;
 
 import com.google.firebase.iid.FirebaseInstanceId;
 import com.google.firebase.iid.FirebaseInstanceIdService;
 import javax.inject.Inject;
 import org.kinecosystem.kinit.KinitApplication;
-import org.kinecosystem.kinit.network.OnboardingService;
+import org.kinecosystem.kinit.server.OnboardingService;
 
 
 public class KinInstanceIdService extends FirebaseInstanceIdService {
