+++ app/src/main/java/org/kinecosystem/kinit/server/ServicesCommonData.kt
@@ -1,8 +1,7 @@
-package org.kinecosystem.kinit.network
+package org.kinecosystem.kinit.server
 
 const val ERROR_NO_INTERNET = 1
 const val ERROR_NO_PUBLIC_ADDRESS = 2
 const val ERROR_EMPTY_RESPONSE = 3
 const val ERROR_APP_SERVER_FAILED_RESPONSE = 4
 const val ERROR_INVALID_DATA = 5
-const val USER_HEADER_KEY = "X-USERID"
