+++ app/src/main/java/org/kinecosystem/kinit/server/ServicesProvider.kt
@@ -1,14 +1,17 @@
-package org.kinecosystem.kinit.network
+package org.kinecosystem.kinit.server
 
 import android.content.Context
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import org.kinecosystem.kinit.BuildConfig
 import org.kinecosystem.kinit.analytics.Analytics
+import org.kinecosystem.kinit.blockchain.Wallet
 import org.kinecosystem.kinit.repository.DataStoreProvider
 import org.kinecosystem.kinit.repository.OffersRepository
 import org.kinecosystem.kinit.repository.TasksRepository
 import org.kinecosystem.kinit.repository.UserRepository
+import org.kinecosystem.kinit.server.api.*
+import org.kinecosystem.kinit.util.NetworkUtils
 import org.kinecosystem.kinit.util.Scheduler
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
@@ -58,16 +61,16 @@ class ServicesProvider {
 
 
         walletService = Wallet(context.applicationContext, dataStoreProvider, userRepo, tasksRepo, analytics,
-            onboardingApi, walletApi, scheduler)
+                onboardingApi, walletApi, scheduler)
         taskService = TaskService(applicationContext,
-            retrofit.create<TasksApi>(TasksApi::class.java),
-            tasksRepo, userRepo.userId(), walletService)
+                retrofit.create<TasksApi>(TasksApi::class.java),
+                tasksRepo, userRepo.userId(), walletService)
         onBoardingService = OnboardingService(applicationContext,
-            retrofit.create<OnboardingApi>(OnboardingApi::class.java),
-            retrofit.create<PhoneAuthenticationApi>(PhoneAuthenticationApi::class.java),
-            userRepo, analytics, taskService, walletService)
+                retrofit.create<OnboardingApi>(OnboardingApi::class.java),
+                retrofit.create<PhoneAuthenticationApi>(PhoneAuthenticationApi::class.java),
+                userRepo, analytics, taskService, walletService)
         offerService = OfferService(applicationContext, retrofit.create<OffersApi>(OffersApi::class.java),
-            userRepo.userId(), offerRepo, analytics, walletService, scheduler)
+                userRepo.userId(), offerRepo, analytics, walletService, scheduler)
     }
 
     fun isNetworkConnected(): Boolean {
