+++ app/src/main/java/org/kinecosystem/kinit/server/api/ApiConsts.kt
@@ -1,0 +1,3 @@
+package org.kinecosystem.kinit.server.api
+
+const val USER_HEADER_KEY = "X-USERID"
