+++ app/src/main/java/org/kinecosystem/kinit/view/adapter/CouponsListAdapter.kt
@@ -10,7 +10,7 @@ import org.kinecosystem.kinit.KinitApplication
 import org.kinecosystem.kinit.R
 import org.kinecosystem.kinit.databinding.CouponRowLayoutBinding
 import org.kinecosystem.kinit.model.spend.Coupon
-import org.kinecosystem.kinit.network.ServicesProvider
+import org.kinecosystem.kinit.server.ServicesProvider
 import org.kinecosystem.kinit.viewmodel.balance.CouponViewModel
 import javax.inject.Inject
