+++ app/src/main/java/org/kinecosystem/kinit/view/phoneVerify/PhoneVerifyActivity.java
@@ -1,6 +1,8 @@
 package org.kinecosystem.kinit.view.phoneVerify;
 
+import android.app.AlertDialog;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 
@@ -8,8 +10,8 @@ import org.kinecosystem.kinit.KinitApplication;
 import org.kinecosystem.kinit.R;
 import org.kinecosystem.kinit.analytics.Analytics;
 import org.kinecosystem.kinit.analytics.Events;
-import org.kinecosystem.kinit.network.OperationCompletionCallback;
 import org.kinecosystem.kinit.repository.UserRepository;
+import org.kinecosystem.kinit.server.OperationCompletionCallback;
 import org.kinecosystem.kinit.util.GeneralUtils;
 import org.kinecosystem.kinit.view.BaseActivity;
 import org.kinecosystem.kinit.view.MainActivity;
@@ -77,10 +79,24 @@ public class PhoneVerifyActivity extends BaseActivity implements PhoneVerificati
     @Override
     public void onSendPhone(String phoneNumber) {
         sendPhoneCount++;
-        model.startPhoneNumberVerification(phoneNumber);
-        getSupportFragmentManager().beginTransaction()
-                .replace(R.id.fragment_container, CodeVerificationFragment.newInstance(phoneNumber, sendPhoneCount >= MAX_SEND_PHONE_COUNT), FRAGMENT_CODE_TAG)
-                .commit();
+        if (model.startPhoneNumberVerification(phoneNumber)) {
+            getSupportFragmentManager().beginTransaction()
+                    .replace(R.id.fragment_container, CodeVerificationFragment.newInstance(phoneNumber, sendPhoneCount >= MAX_SEND_PHONE_COUNT), FRAGMENT_CODE_TAG)
+                    .commit();
+        } else {
+            showDialogBlackedList();
+        }
+    }
+
+    private void showDialogBlackedList() {
+        new AlertDialog.Builder(this, R.style.CustomAlertDialog).setTitle(getResources().getString(R.string.oh_no)).setMessage(getResources().getString(R.string.block_area_code_message))
+                .setPositiveButton(getResources().getString(R.string.dialog_ok), new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialogInterface, int i) {
+                        dialogInterface.dismiss();
+                        finish();
+                    }
+                }).create().show();
     }
 
     @Override
