+++ app/src/test/java/org/kinecosystem/kinit/mock/MockComponentsProvider.kt
@@ -1,9 +1,9 @@
 package org.kinecosystem.kinit.mock
 
 import org.kinecosystem.kinit.analytics.Analytics
-import org.kinecosystem.kinit.network.ServicesProvider
-import org.kinecosystem.kinit.network.TaskService
-import org.kinecosystem.kinit.network.Wallet
+import org.kinecosystem.kinit.server.ServicesProvider
+import org.kinecosystem.kinit.server.TaskService
+import org.kinecosystem.kinit.blockchain.Wallet
 import org.kinecosystem.kinit.notification.NotificationPublisher
 import org.kinecosystem.kinit.repository.*
 import org.mockito.Mockito
@@ -17,7 +17,7 @@ class MockComponentsProvider : DataStoreProvider {
     var analytics: org.kinecosystem.kinit.analytics.Analytics = Mockito.mock(Analytics::class.java)
     var scheduler = MockScheduler()
     var wallet: Wallet = Mockito.mock(Wallet::class.java)
-    val taskService:TaskService = Mockito.mock(TaskService::class.java)
+    val taskService: TaskService = Mockito.mock(TaskService::class.java)
     var notificationPublisher: NotificationPublisher = Mockito.mock(NotificationPublisher::class.java)
